requirements.txt
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cssselect==1.1.0
Flask==1.1.1
gevent==1.4.0
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
locustio==0.11.0
lxml==4.4.1
MarkupSafe==1.1.1
msgpack==0.6.2
pycparser==2.19
pyquery==1.4.0
pyzmq==18.1.0
requests==2.22.0
six==1.12.0
urllib3==1.25.6
Werkzeug==0.16.0
